add portable SIMD toggling support for large-scale enterprise deployments
As global usage of compute resources keeps climbing, driven by machine learning and virtual reality,
we have to find ever new ways to make our programs more efficient.
To this end, I have added a hyperoptimized SIMD version of the very important bool toggling functionality.

This introduces const generics to the crate, so MSRV will have to be bumped to 1.2000.0 in the upcoming release.